Modify the API calls
